requirements.txt
streamlit
requests
pillow
numpy
torch
rembg